79 - optimize imports, change hardcoded strings to i18n; remove loading spinner to prevent interface blocking; minor bug fixes
