+++ src/scenes/CityChoice.js
@@ -1,14 +1,11 @@
 import React, { Component, PropTypes } from 'react';
-
 import { AsyncStorage, View, StyleSheet, Image } from 'react-native';
 import { connect } from 'react-redux';
 import Spinner from 'react-native-loading-spinner-overlay';
-
 import LocationListView from '../components/LocationListView';
 import ApiClient from '../utils/ApiClient';
 import I18n from '../constants/Messages';
-
-const styles = require('../styles');
+import styles from '../styles';
 
 class CityChoice extends Component {
