+++ src/utils/DrawerCommons.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import { AsyncStorage, Platform } from 'react-native';
 import { Drawer } from 'react-native-material-design';
-
 import I18n from '../constants/Messages';
 
 export default class DrawerCommons {
