+++ src/utils/ServiceCommons.js
@@ -6,28 +6,8 @@ import {
     StyleSheet
 } from 'react-native';
 import { default as Icon } from 'react-native-vector-icons/FontAwesome';
-
 import I18n from '../constants/Messages';
-
-const styles = StyleSheet.create({
-    icon: {
-        width: 32,
-        height: 32
-    },
-    container: {
-        flex: 1,
-        flexDirection: 'row'
-    },
-    iconContainer: {
-        width: 32,
-        marginRight: 10,
-        flexDirection: 'row',
-        alignItems: 'center'
-    },
-    detailsContainer: {
-        flexDirection: 'column'
-    }
-});
+import styles from '../styles';
 
 export default class ServiceCommons {
 
@@ -46,14 +26,14 @@ export default class ServiceCommons {
         let stars = this.renderStars(service.rating);
 
         return (
-            <View style={styles.container}>
+            <View style={styles.rowContainer}>
                 <View style={styles.iconContainer}>
                     <Image
                         source={{uri: serviceType.icon_url}}
-                        style={styles.icon}
+                        style={styles.mapIcon}
                     />
                 </View>
-                <View style={styles.detailsContainer}>
+                <View>
                     <Text>{service.name}</Text>
                     <Text>{I18n.t('RATING')}: {stars}</Text>
                     <Text>{locationName}</Text>
